Move image processor and resolver to tce-image 🎨
