+++ client/components/content-elements/tce-image/index.js
@@ -1,12 +1,11 @@
 import Edit from './edit';
+import info from './info';
 import Toolbar from './edit/Toolbar';
 
 const initState = () => ({ url: null });
 
 export default {
-  name: 'Image',
-  type: 'IMAGE',
-  version: '1.0',
+  ...info,
   initState,
   Edit,
   Toolbar,
