+++ client/components/content-elements/tce-image/info.js
@@ -1,0 +1,5 @@
+module.exports = {
+  name: 'Image',
+  type: 'IMAGE',
+  version: '1.0'
+};
